move to separate class default gh-push-event listener
add javadocs
add simple test for filtering by trigger
