+++ src/main/java/org/jenkinsci/plugins/github/webhook/listener/DefaultPushGHEventListener.java
@@ -1,0 +1,41 @@
+package org.jenkinsci.plugins.github.webhook.listener;
+
+import com.cloudbees.jenkins.GitHubPushTrigger;
+import hudson.Extension;
+import hudson.model.AbstractProject;
+import org.jenkinsci.plugins.github.webhook.GHEventsListener;
+import org.kohsuke.github.GHEvent;
+
+import java.util.Set;
+
+import static com.google.common.collect.Sets.immutableEnumSet;
+import static org.jenkinsci.plugins.github.util.JobInfoHelpers.withTrigger;
+import static org.kohsuke.github.GHEvent.PUSH;
+
+/**
+ * @author lanwen (Merkushev Kirill)
+ * @since 1.11.4
+ */
+@Extension
+@SuppressWarnings("unused")
+public class DefaultPushGHEventListener extends GHEventsListener {
+    /**
+     * This listener is applicable only for job with GHPush trigger
+     *
+     * @param project to check for trigger
+     *
+     * @return true if project has {@link GitHubPushTrigger}
+     */
+    @Override
+    public boolean isApplicable(AbstractProject<?, ?> project) {
+        return withTrigger(GitHubPushTrigger.class).apply(project);
+    }
+
+    /**
+     * @return set with only push event
+     */
+    @Override
+    public Set<GHEvent> events() {
+        return immutableEnumSet(PUSH);
+    }
+}
